Fix out of sync comments issue
The threads variable was outdated when a newComment was added to it, causing the list to always be missing the new comments added after the fetch. By using the function form of setState, the state is guaranteed to be up to date as it doesn't rely on a re-render to have up to date state.